+++ src/main/java/ru/clevertec/core/service/domain/ConverterMapNode.java
@@ -1,10 +1,11 @@
-package ru.clevertec.core;
+package ru.clevertec.core.service.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
+import ru.clevertec.core.ContainerCreator;
+import ru.clevertec.core.ContainerData;
 import ru.clevertec.core.node.Node;
 import ru.clevertec.core.node.ObjectNode;
-import ru.clevertec.core.node.ValueNode;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
