+++ src/main/java/ru/clevertec/core/service/json/ConverterObjectAsString.java
@@ -1,4 +1,4 @@
-package ru.clevertec.core.service;
+package ru.clevertec.core.service.json;
 
 public class ConverterObjectAsString implements ConverterToJson{
